studentPage code (part1)
We wrote the code with the database in mind. In theory it should be plug and play(as long as its correct). Type in your file name in line 11's URL variable and it should send an output of CAMPUS,MAJOR, and TERM to said URL when the submit button is pressed.