StudentPageCode Part2
we want the admin to be able to add a CAMPUS,MAJOR, and TERM to the database.  This way we can just have our webpage read your database for those fields and add them automatically to our dropdowns, rather than us continuously having to hard code it into the html file. We'd like it (if possible) to have a way to get the values of all domains for CAMPUS,MAJOR,and TERM in your database in an array/list/etc form.