Files to delete Map Uploaded
The code searches for maps and displays them in a semester table properly. I wrote a php file which is supposed to delete the map but was having trouble with it. I don't want to waste our time on it, as Im a bit slow with it. Whats left is to style it and connect it to the php file. Have at it team.